backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0
pydantic[email]
python-dotenv==1.0.0
requests==2.31.0
langchain==0.1.0
langgraph==0.0.26
langchain-community==0.0.10
networkx==3.2.1
numpy==1.26.2
pandas==2.1.4
python-docx==1.1.0
docxtpl==0.16.7
weasyprint==60.1
